requirements.txt
python-telegram-bot==22.0
gspread==6.1.2
oauth2client==4.1.3
python-dotenv==1.0.1
httpx==0.28.1